+++ debug/jt/jt-808-server-spring-boot-starter-reactive-debug/src/main/java/io/github/hylexus/xtream/debug/ext/jt808/Jt808SpringBootStarterDebugApp.java
@@ -16,7 +16,6 @@
 
 package io.github.hylexus.xtream.debug.ext.jt808;
 
-import io.netty.util.ResourceLeakDetector;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
@@ -30,7 +29,7 @@ public class Jt808SpringBootStarterDebugApp {
      */
     public static void main(String[] args) {
         // FIXME 如果你不了解 ResourceLeakDetector 是做什么的,请务必注释掉下面这行代码!!!
-        ResourceLeakDetector.setLevel(ResourceLeakDetector.Level.PARANOID);
+        // io.netty.util.ResourceLeakDetector.setLevel(io.netty.util.ResourceLeakDetector.Level.PARANOID);
         SpringApplication.run(Jt808SpringBootStarterDebugApp.class, args);
     }
 }